Reorganize categories code
Seems that the list of categories belongs with the services, while the
support code belongs with base service.

Ref #2832
