+++ core/base-service/loader.js
@@ -2,7 +2,7 @@
 
 const path = require('path')
 const glob = require('glob')
-const { categories } = require('../../services/categories')
+const categories = require('../../services/categories')
 const BaseService = require('./base')
 const { assertValidServiceDefinitionExport } = require('./service-definitions')
