+++ services/categories.js
@@ -1,8 +1,6 @@
 'use strict'
 
-const Joi = require('joi')
-
-const categories = [
+module.exports = [
   { id: 'build', name: 'Build' },
   { id: 'coverage', name: 'Code Coverage' },
   { id: 'analysis', name: 'Analysis' },
@@ -21,18 +19,3 @@ const categories = [
   { id: 'activity', name: 'Activity' },
   { id: 'other', name: 'Other' },
 ]
-
-const isRealCategory = Joi.equal(categories.map(({ id }) => id)).required()
-
-const isValidCategory = Joi.alternatives()
-  .try(isRealCategory, Joi.equal('debug', 'dynamic').required())
-  .required()
-
-function assertValidCategory(category, message = undefined) {
-  Joi.assert(category, isValidCategory, message)
-}
-
-module.exports = {
-  categories,
-  assertValidCategory,
-}
